% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chow.test.R
\name{chow.test}
\alias{chow.test}
\title{Chow Test for Structural Break}
\usage{
chow.test(Y, SB, nboot = 500, lags = 12, start = NULL, end = NULL,
  frequency = NULL, format = NULL, dateVector = NULL)
}
\arguments{
\item{Y}{Data}

\item{SB}{integer. Structural break either of type integer as the number of observations which belong to the pre-break period or
Date character. If a date character is provided, either a date Vector which contains the time line of the data
in corresponding format or then the conventional time parameters need to be provided.}

\item{nboot}{??}

\item{lags}{Maximum number of Lag order}

\item{start}{character. Start of the time series (only if dateVector is empty)}

\item{end}{character. End of the time series (only if dateVector is empty)}

\item{frequency}{character. Frequency of the time series (only if dateVector is empty)}

\item{format}{character. Date format (only if dateVector is empty)}

\item{dateVector}{vector. Vector of the time period concerned containing SB}
}
\value{
A list of results
}
\description{
The Chow test is applied to a time series an presupposed point of structural break.
}
