% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.dc.R
\name{id.dc}
\alias{id.dc}
\title{Least dependent innovations identification based on distance covariance matrix}
\usage{
id.dc(x)
}
\arguments{
\item{x}{VAR-object. (S)VAR model to be identified}
}
\value{
A list of class "svars" with elements
\item{B}{Estimated instanstaneous response matrix B, i.e. unique decomposition of the covariance matrix of reduced form errors}
\item{A_hat}{Estimated VAR parameter}
\item{method}{Method applied for identifaction}
\item{n}{Number of observations}
\item{type}{Type of the VAR model, e.g. 'const'}
}
\description{
Identify the instantaneous response matrix B in SVAR models based on least dependent innovations.
}
\examples{
\dontrun{
# data contains quarterly observations from 1965Q1 to 2008Q3
# x = output gap
# pi = inflation
# i = interest rates
set.seed(23211)
v1 <- VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.dc(v1)
summary(x1)

# switching columns according to sign pattern
x1$B <- x1$B[,c(3,2,1)]
x1$B[,3] <- x1$B[,3]*(-1)

# impulse response analysis
i1 <- imrf(x1, horizon = 30)
plot(i1, scales = 'free_y')
}

}
