% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.iv.R
\name{id.iv}
\alias{id.iv}
\title{Instrument variable identification of SVAR}
\usage{
id.iv(x, instruments = NULL)
}
\arguments{
\item{x}{An object of class 'vars', 'vec2var', 'nlVar'. Estimated VAR object}

\item{PIT}{Logical. If PIT='TRUE', the distribution and density of the independent components are estimated using gaussian kernel density estimates}
}
\value{
A list of class "svars" with elements
\item{B}{Estimated structural impact matrix B, i.e. unique decomposition of the covariance matrix of reduced form errors}
\item{A_hat}{Estimated VAR parameter}
\item{method}{Method applied for identification}
\item{n}{Number of observations}
\item{type}{Type of the VAR model, e.g. 'const'}
\item{y}{Data matrix}
\item{p}{Number of lags}
\item{K}{Dimension of the VAR}
\item{F_statistic}{Test statistic for weak instrument}
}
\description{
Given an estimated VAR model, this function applies independence-based identification for the structural impact matrix B of the corresponding SVAR model
\deqn{y_t=c_t+A_1 y_{t-1}+...+A_p y_{t-p}+u_t  =c_t+A_1 y_{t-1}+...+A_p y_{t-p}+B \epsilon_t.}
Matrix B corresponds to the unique decomposition of the least squares covariance matrix \eqn{\Sigma_u=B B'} if the vector of structural shocks \eqn{\epsilon_t} contains at most one Gaussian shock (Comon, 1994).
A nonparametric dependence measure, the distance covariance (Szekely et al, 2007), determines least dependent structural shocks. The algorithm described in Matteson and Tsay (2013) is applied to calculate the matrix B.
}
\examples{


}
\references{
Lunsford, K. G., 2016. "Identifying Structural VARs with a Proxy Variable and a Test for a Weak Proxy"
}
\seealso{
For alternative identification approaches see \code{\link{id.st}}, \code{\link{id.cvm}}, \code{\link{id.cv}} or \code{\link{id.ngml}}
}
