% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.svars.R
\name{irf}
\alias{irf}
\alias{irf.svars}
\title{Impulse Response Functions for SVAR Models}
\usage{
\method{irf}{svars}(x, n.ahead = 20, ...)
}
\arguments{
\item{x}{SVAR object of class "svars"}

\item{n.ahead}{Integer specifying the steps.}

\item{...}{Currently not used.}
}
\description{
Calculation of impulse response functions for an identified SVAR object 'svars' derived by function id.cvm( ),id.cv( ),id.dc( ), id.ngml( ) or id.st( ).
}
\examples{
\donttest{
v1 <- vars::VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.ngml(v1)
x2 <- irf(x1, n.ahead = 20)
plot(x2)
}

}
