\name{svarIdent-package}
\alias{svarIdent-package}
\alias{svarIdent}
\docType{package}
\title{
\packageTitle{svarIdent}
}
\description{
\packageDescription{Implements data driven identification methods for SVAR models. Based on an existing VAR model objct the B matrix is obtained via changes in volatility, least dependent innovations or non gaussian maximum likelihood.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{svarIdent}
\packageIndices{svarIdent}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Alexander Lange and Bernhard Dalheimer}

Maintainer: \packageMaintainer{Alexander Lange <alexander.lange@stud.uni-goettingen.de>}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{SVAR, VAR}

}
\examples{
~~ simple examples of the most important functions ~~
}
