\name{USA}

\docType{data}

\alias{USA}

\title{USA macroeconomic time series}

\description{
  \cr
  Quarterly observations from 1965Q1 to 2008Q3
  \tabular{ll}{
    x \tab Output gap \cr
    pi \tab Inflation \cr
    i \tab Interest rate \cr
  }

}

\usage{USA}

\format{A \code{data.frame} containing 174 observations on 3 variables.}

\source{??}

