% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{wild.boot}
\alias{wild.boot}
\title{Bootstrapping procedure}
\usage{
wild.boot(x, radermacher = FALSE, horizon, nboot)
}
\arguments{
\item{x}{SVAR object of class "svars"}

\item{horizon}{Number of observations in time to be included in}

\item{nboot}{Number of bootstrap iterations}
}
\description{
Calculating confidance bands for impulse response via wild bootsrap
}
