% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/id.cv.R
\name{id.cv}
\alias{id.cv}
\title{Changes in Volatility Identification}
\usage{
id.cv(x, SB, start = NULL, end = NULL, frequency = NULL, format = NULL,
  dateVector = NULL)
}
\arguments{
\item{x}{VAR-object. (S)VAR model to determine B matrix for}

\item{SB}{integer. Structural break either of type integer as the number of observations which belong to the pre-break period or
Date character. If a date character is provided, either a date Vector which contains the time line of the data
in corresponding format or then the conventional time parameters need to be provided.}

\item{start}{character. Start of the time series (only if dateVector is empty)}

\item{end}{character. End of the time series (only if dateVector is empty)}

\item{frequency}{character. Frequency of the time series (only if dateVector is empty)}

\item{format}{character. Date format (only if dateVector is empty)}

\item{dateVector}{vector. Vector of the time period concerned containing SB}
}
\value{
A list of results
}
\description{
Identify B matrix based on changes in Volatility.
}
\examples{
require(vars)
# data contains quartlery observations from 1965Q1 to 2008Q3
# assumed structural break in 1979Q4
# x = output gap
# pi = inflation
# i = interest rates

set.seed(23211)
v1 <- VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.cv(v1, SB = 60)
}

