% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_ngml.R
\name{id.ngml}
\alias{id.ngml}
\title{non-Gaussian maximum likelihood Identification}
\usage{
id.ngml(x)
}
\arguments{
\item{x}{VAR-object. (S)VAR model to determine B matrix for}
}
\value{
A list of class "svarIdent" with elements
\item{B}{Estimated B matrix, i.e. unique decomposition of the covariance matrix}
\item{sigma}{Estimated scale of the standardized B matrix}
\item{sigma_SE}{Standard errors of the scale}
\item{df}{Estimated degrees freedom}
\item{df_SE}{Standard errors of the degrees of freedom}
\item{Fish}{Observerd Fisher information matrix}
\item{A_hat}{Estimated VAR parameter}
\item{B_stand}{Estimated standardized B matrix}
\item{B_stand_SE}{Standard errors of standardized B matrix}
\item{Lik}{Function value of likelihood}
\item{method}{The applied identifaction method}
\item{n}{Number of observations}
\item{type}{Type of the VAR model, e.g. 'const'}
}
\description{
Identify B matrix based on non-Gaussian maximum likelihood.
}
\examples{
\dontrun{
require(vars)
require(ggplot2)
require(reshape2)
# data contains quartlery observations from 1965Q1 to 2008Q3
# assumed structural break in 1979Q4
# x = output gap
# pi = inflation
# i = interest rates
set.seed(23211)
v1 <- VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.mgml(v1)
summary(x1)

x1$B <- x1$B[,c(3,2,1)]
x1$B[,3] <- x1$B[,3]*(-1)

plot(x1, horizon = 30, scales = 'free_y')
}

}
