% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/id_ngml.R
\name{id.ngml}
\alias{id.ngml}
\title{non-Gaussian maximum likelihood Identification}
\usage{
id.ngml(x, stage3 = FALSE)
}
\arguments{
\item{x}{VAR-object. (S)VAR model to determine B matrix for}

\item{stage3}{Logical. Whether the VAR parameter should be estimated via non-gaussian maximum likelihood (computationally demanding)}
}
\value{
A list of class "svars" with elements
\item{B}{Estimated B matrix, i.e. unique decomposition of the covariance matrix}
\item{sigma}{Estimated scale of the standardized B matrix}
\item{sigma_SE}{Standard errors of the scale}
\item{df}{Estimated degrees freedom}
\item{df_SE}{Standard errors of the degrees of freedom}
\item{Fish}{Observerd Fisher information matrix}
\item{A_hat}{Estimated VAR parameter}
\item{B_stand}{Estimated standardized B matrix}
\item{B_stand_SE}{Standard errors of standardized B matrix}
\item{Lik}{Function value of likelihood}
\item{method}{The applied identifaction method}
\item{n}{Number of observations}
\item{type}{Type of the VAR model, e.g. 'const'}
}
\description{
Identify B matrix based on non-Gaussian maximum likelihood.
}
\examples{
\dontrun{
require(vars)
require(ggplot2)
require(reshape2)
# data contains quartlery observations from 1965Q1 to 2008Q3
# x = output gap
# pi = inflation
# i = interest rates
set.seed(23211)
v1 <- VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.ngml(v1)
summary(x1)

# switching columns according to sign patter
x1$B <- x1$B[,c(3,2,1)]
x1$B[,3] <- x1$B[,3]*(-1)

# Impulse response Analysis
i1 <- irf(x1, horizon = 30)
plot(i1, scales = 'free_y')
}
}

