% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.ngml.R
\name{id.ngml}
\alias{id.ngml}
\title{Non-Gaussian maximum likelihood identification of SVAR models}
\usage{
id.ngml(x, stage3 = FALSE)
}
\arguments{
\item{x}{An object of class 'vars', 'vec2var', 'nlVar'. Estimated VAR object}

\item{stage3}{Logical. If stage3="TRUE", the VAR parameters are estimated via non-gaussian maximum likelihood (computationally demanding)}
}
\value{
A list of class "svars" with elements
\item{B}{Estimated structural impact matrix B, i.e. unique decomposition of the covariance matrix of reduced form errors}
\item{sigma}{Estimated scale of the standardized matrix B_stand, i.e. \eqn{B=B_stand*diag(\sigma_1,...,\sigma_K)}}
\item{sigma_SE}{Standard errors of the scale}
\item{df}{Estimated degrees of freedom}
\item{df_SE}{Standard errors of the degrees of freedom}
\item{Fish}{Observed Fisher information matrix}
\item{A_hat}{Estimated VAR parameter}
\item{B_stand}{Estimated standardized structural impact matrix}
\item{B_stand_SE}{Standard errors of standardized matrix B_stand}
\item{Lik}{Function value of likelihood}
\item{method}{Method applied for identifaction}
\item{n}{Number of observations}
\item{type}{Type of the VAR model, e.g. 'const'}
}
\description{
Given an estimated VAR model, this function applies identification by means of a non-Gaussian likelihood for the structural impact matrix B of the corresponding SVAR model
\deqn{y_t=c_t+A_1 y_{t-1}+...+A_p y_{t-p}+u_t   =c_t+A_1 y_{t-1}+...+A_p y_{t-p}+B \epsilon_t.}
Matrix B corresponds to the unique decomposition of the least squares covariance matrix \eqn{\Sigma_u=B B'} if the vector of structural shocks \eqn{\epsilon_t} contains at most one Gaussian shock (Comon, 94).
A likelihood function of independent t-distributed structural shocks \eqn{\epsilon_t=B^{-1}u_t} is maximized with respect to the entries of B and the degrees of freedom of the t-distribution (Lanne et al., 2017).
}
\examples{
\donttest{
# data contains quarterly observations from 1965Q1 to 2008Q3
# x = output gap
# pi = inflation
# i = interest rates
set.seed(23211)
v1 <- VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.ngml(v1)
summary(x1)

# switching columns according to sign pattern
x1$B <- x1$B[,c(3,2,1)]
x1$B[,3] <- x1$B[,3]*(-1)

# impulse response analysis
i1 <- imrf(x1, horizon = 30)
plot(i1, scales = 'free_y')
}
}
\references{
Lanne, M., Meitz, M., Saikkonen, P., 2017. Identification and estimation of non-Gaussian structural vector autoregressions. J. Econometrics 196 (2), 288-304.\cr
Comon, P., 1994. Independent component analysis, A new concept?, Signal Processing, 36, 287-314
}
\seealso{
For alternative identification approaches see \code{\link{id.cvm}}, \code{\link{id.dc}} or \code{\link{id.cv}}
}
