% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.st.R
\name{id.st}
\alias{id.st}
\title{Identification of SVAR models by means of a smooth transition of volatility}
\usage{
id.st(x, nc = 1, c_lower = 0.3, c_upper = 0.7, c_step = 5,
  gamma_lower = -3, gamma_upper = 2, gamma_step = 0.5, max.iter = 50,
  crit = 0.01)
}
\arguments{
\item{x}{An object of class 'vars', 'vec2var', 'nlVar'. Estimated VAR object}

\item{nc}{Number of processor cores (Not available on windows machines).
Note the smooth transition model is computationally extremely demanding.}

\item{c_lower}{Start point where the algorithm starts to search for the volatility shift.
Defalt is 0.3*(Total number of observations)}

\item{c_upper}{End point where the algorithm stops to search for the volatility shift.
Defalt is 0.7*(Total number of observations)}

\item{c_step}{Step width of c, default is 5.}

\item{gamma_lower}{Lower bound for gamma. Small values indicate a flat transition function. Default is -3}

\item{gamma_upper}{Upper bound for gamma. Large values indicate a steep transition function. Default is 2}

\item{gamma_step}{Step width of gamma, default is 0.5}

\item{max.iter}{Integer. Number of maximum GLS iterations}

\item{crit}{Integer. Critical value for the precision of the GLS estimation}
}
\value{
A list of class "svars" with elements
\item{Lambda}{Estimated heteroscedasticity matrix \eqn{\Lambda}}
\item{Lambda_SE}{Matrix of standard errors of Lambda}
\item{B}{Estimated structural impact matrix B, i.e. unique decomposition of the covariance matrix of reduced form residuals}
\item{B_SE}{Standard errors of matrix B}
\item{n}{Number of observations}
\item{Fish}{Observed Fisher information matrix}
\item{Lik}{Function value of likelihood}
\item{wald_statistic}{Results of pairwise Wald tests}
\item{iteration}{Number of GLS estimations}
\item{method}{Method applied for identification}
\item{SB}{Structural break (number of observations)}
\item{SBcharacter}{Structural break (date; if provided in function arguments)}
}
\description{
Given an estimated VAR model, this function identifies the structural impact matrix B of the corresponding SVAR model
by means of smooth transition
\deqn{y_t=c_t+A_1 y_{t-1}+...+A_p y_{t-p}+u_t
=c_t+A_1 y_{t-1}+...+A_p y_{t-p}+B \epsilon_t.}
Matrix B corresponds to the decomposition of the pre-break covariance matrix \eqn{\Sigma_1=B B'}.
The post-break covariance corresponds to \eqn{\Sigma_2=B\Lambda B'} where \eqn{\Lambda} is the estimated unconditional heteroskedasticity matrix.
}
\examples{
\donttest{
# data contains quartlery observations from 1965Q1 to 2008Q2
# assumed structural break in 1979Q4
# x = output gap
# pi = inflation
# i = interest rates
set.seed(23211)
v1 <- vars::VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.st(v1,)
summary(x1)

# switching columns according to sign patter
x1$B <- x1$B[,c(3,2,1)]
x1$B[,3] <- x1$B[,3]*(-1)

# Impulse response analysis
i1 <- imrf(x1, horizon = 30)
plot(i1, scales = 'free_y')

# Restrictions
# Assuming that the interest rate doesn't influence the output gap on impact
restMat <- matrix(rep(NA, 9), ncol = 3)
restMat[1,3] <- 0
x2 <- id.cv(v1, SB = 60, restriction_matrix = restMat)

#Structural brake via Dates
# given that time series vector with dates is available
dateVector = seq(as.Date("1965/1/1"), as.Date("2008/7/1"), "quarter")
x3 <- id.cv(v1, SB = "1985-01-01", format = "\%Y-\%m-\%d", dateVector = dateVector)

# or pass sequence arguments directly
x4 <- id.cv(v1, SB = "1985-01-01", format = "\%Y-\%m-\%d", start = "1965-01-01", end = "2008-06-01",
frequency = "quarter")

# or provide ts date format (For quarterly, monthly, weekly and daily frequencies only)
x5 <- id.cv(v1, SB = c(1985, 1))

}
}
\references{
Rigobon, R., 2003. Identification through Heteroskedasticity. The Review of Economics and Statistics, 85, 777-792.\cr
 Herwartz, H. & Ploedt, M., 2016. Simulation Evidence on Theory-based and Statistical Identification under Volatility Breaks Oxford Bulletin of Economics and Statistics, 78, 94-112.
}
\seealso{
For alternative identification approaches see \code{\link{id.cvm}}, \code{\link{id.dc}} or \code{\link{id.ngml}}
}
