% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse Response Function}
\usage{
irf(x, horizon = 20)
}
\arguments{
\item{x}{SVAR object of class "svarIdent"}

\item{horizon}{Number of observations in time to be included in}
}
\description{
Calculation of impulse response function for a given identified SVAR object
}
\examples{
\dontrun{

x2 <- irf(x1, horizon = 20)
plot(x2)
}

}
