\name{USA}

\docType{data}

\alias{USA}

\title{USA macroeconomic time series}

\description{
  \cr
  Quarterly observations from 1965Q1 to 2008Q3
  \tabular{ll}{
    x \tab Output gap \cr
    pi \tab Inflation \cr
    i \tab Interest rate \cr
  }

}

\usage{USA}

\format{A \code{data.frame} containing 174 observations on 3 variables.}

\source{Herwartz, H. & Plödt, M., 2016. Simulation Evidence on Theory-based and Statistical Identification under Volatility Breaks, Oxford Bulletin of Economics and Statistics, 78, 94-112. data originally from FRED database of the Federal Reserve Bank of St. Louis}

