% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_wild.R
\name{wild.boot}
\alias{wild.boot}
\title{Bootstrap for IRFs of identified SVARs}
\usage{
wild.boot(x, radermacher = FALSE, horizon, nboot, nc)
}
\arguments{
\item{x}{SVAR object of class "svars"}

\item{radermacher}{If radermacher="TRUE", the radermacher distribution is used to generate the bootstrap samples}

\item{horizon}{Time horizon of impulse response functions}

\item{nboot}{Number of bootstrap iterations}

\item{nc}{Number of processor cores (Not available on windows machines)}
}
\description{
Calculating confidance bands for impulse response functions via wild bootsrap techniques
}
\examples{
\dontrun{
# data contains quarterly observations from 1965Q1 to 2008Q3
# x = output gap
# pi = inflation
# i = interest rates
set.seed(23211)
v1 <- VAR(USA, lag.max = 10, ic = "AIC" )
x1 <- id.ngml(v1)
summary(x1)

# switching columns according to sign patter
x1$B <- x1$B[,c(3,2,1)]
x1$B[,3] <- x1$B[,3]*(-1)

# impulse response analysis with confidence bands
bb <- wild.boot(x1, radermacher = T, nboot = 100, horizon = 30)
plot(bb, lowerq = 0.16, upperq = 0.84)
}


}
